[Infra] Core 모듈 생성, 코드 분리
