+++ weave-iOS/Projects/App/Sources/MyPage/Profile/Feature/MyPageFeature.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import ComposableArchitecture
 import Services
 import AVFoundation
+import DesignSystem
 
 struct MyPageFeature: Reducer {
     struct State: Equatable {
