+++ weave-iOS/Projects/App/Sources/SignUp/View/StepSubViews/SignUpUniversityView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 import ComposableArchitecture
 import DesignSystem
+import CoreKit
 
 struct SignUpUniversityView: View {
