+++ weave-iOS/Projects/App/Sources/SignUp/View/StepSubViews/SignUpYearView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 import ComposableArchitecture
 import DesignSystem
+import CoreKit
 
 struct SignUpYearView: View {
     let store: StoreOf<SignUpFeature>
